+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `NoSemiAfterFieldDeclarationRule` analysis rule, which flags field declarations without trailing
   semicolons.
 - `TrailingWhitespaceRule` analysis rule, which flags source code lines with trailing whitespace.
+- `IfThenShortCircuitRule` analysis rule, which flags erroneous `IfThen` usages that may cause
+  access violations.
 
 ### Changed