+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2020-08-07
+
 ### Added
 
 - Support for pascal-style result assignments in semantic analysis.
@@ -556,7 +558,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.22.1...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.23.0...HEAD
+[0.23.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.22.1...v0.23.0
 [0.22.1]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.22.0...v0.22.1
 [0.22.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.20.0...v0.21.0